fizing uninstall script for tmux

+++ uninstall.sh
@@ -4,19 +4,21 @@ cd $HOME
 
 rm -rf $HOME/.oh-my-zsh
 rm -rf $HOME/.rcfiles
-rm $HOME/bin
-rm $HOME/.gemrc
-rm $HOME/.gitattributes
-rm $HOME/.gitconfig
-rm $HOME/.gitignore_global
-rm $HOME/.irbeditorrc
-rm $HOME/.irbrc
-rm $HOME/.pryrc
-rm $HOME/.railsrc
-rm $HOME/.rdebugrc
-rm $HOME/.tm_properties
-rm $HOME/.vimrc
-rm $HOME/.zshrc
-rm $HOME/.starship
+
+[[ -L $HOME/.gemrc ]]             && rm $HOME/.gemrc
+[[ -L $HOME/.gitattributes ]]     && rm $HOME/.gitattributes
+[[ -L $HOME/.gitconfig ]]         && rm $HOME/.gitconfig
+[[ -L $HOME/.gitignore_global ]]  && rm $HOME/.gitignore_global
+[[ -L $HOME/.irbeditorrc ]]       && rm $HOME/.irbeditorrc
+[[ -L $HOME/.irbrc ]]             && rm $HOME/.irbrc
+[[ -L $HOME/.pryrc ]]             && rm $HOME/.pryrc
+[[ -L $HOME/.railsrc ]]           && rm $HOME/.railsrc
+[[ -L $HOME/.rdebugrc ]]          && rm $HOME/.rdebugrc
+[[ -L $HOME/.starship ]]          && rm $HOME/.starship
+[[ -L $HOME/.tm_properties ]]     && rm $HOME/.tm_properties
+[[ -L $HOME/.tmux ]]              && rm $HOME/.tmux.conf
+[[ -L $HOME/.vimrc ]]             && rm $HOME/.vimrc
+[[ -L $HOME/.zshrc ]]             && rm $HOME/.zshrc
+[[ -L $HOME/bin ]]                && rm $HOME/bin
 
 mv .zshrc_original .zshrc
